Fix a bug in tmrs_clrtimer which forget to break out the loop after del timer.

+++ kernel/hal/hal.c
@@ -285,5 +285,5 @@ void init_tss(struct cpu *c)
 
 void set_kernel_stack(void *stack)
 {
-	CURR_CPU->arch.tss.esp0 = stack;
+	CURR_CPU->arch.tss.esp0 = (uint32_t)stack;
 }
